Workaround for: Failed to fetch stats from backend: redis: nil
Before any blocks are found I was getting "Failed to fetch stats from backend: redis: nil" and the web UI would not render anything.

This is a work around for this issue that happens when redis returns redis.Nil due to a key not found error in CollectStats.

A similar issue was reported upstream: https://github.com/sammy007/open-ethereum-pool/issues/327

Watching that issue to see how it is resolved.